requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.0
